% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_path.R
\name{get_path}
\alias{get_path}
\title{Get value from any API path}
\usage{
get_path(path, user = getOption(x = "ISO_NE_USER"), password = getOption(x =
  "ISO_NE_PASSWORD"), ...)
}
\arguments{
\item{user}{user account}

\item{password}{password}

\item{...}{passed to httr functions.}

\item{out.tz}{output timezone.}
}
\description{
May need to add \code{config = httr::config(ssl.verifypeer = FALSE)} if SSL cert can't be found.
}

