% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAllHourlySystemLoadForecastForDay.R
\name{getAllHourlySystemLoadForecastForDay}
\alias{getAllHourlySystemLoadForecastForDay}
\title{Get All Hourly System Load Forecast For Day}
\usage{
getAllHourlySystemLoadForecastForDay(day = Sys.Date(), user = getOption(x =
  "ISO_NE_USER"), password = getOption(x = "ISO_NE_PASSWORD"),
  out.tz = "America/New_York", ...)
}
\arguments{
\item{day}{date in \code{'\%Y-\%m-\%d'}, or object coercible to class \code{Date}}

\item{user}{user account}

\item{password}{password}

\item{out.tz}{output timezone.}

\item{...}{passed to httr functions.}
}
\description{
Get all hourly load forecasts for a given day. May need to add \code{config = httr::config(ssl.verifypeer = FALSE)} if SSL cert cant be found.
}

