% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCurrentMorningReport.R
\name{getCurrentMorningReport}
\alias{getCurrentMorningReport}
\title{Get Current Morning Report}
\usage{
getCurrentMorningReport(user = getOption(x = "ISO_NE_USER"),
  password = getOption(x = "ISO_NE_PASSWORD"), out.tz = "America/New_York",
  ...)
}
\arguments{
\item{user}{user account}

\item{password}{password}

\item{out.tz}{output timezone.}

\item{...}{passed to httr functions.}
}
\description{
May need to add \code{config = httr::config(ssl.verifypeer = FALSE)} if SSL cert cant be found.
}

