% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHistorical5MinLMPs.R
\name{getFiveMinuteLMPsAtLocation}
\alias{getFiveMinuteLMPsAtLocation}
\title{Get historical 5-minute LMPs for a date range}
\usage{
getFiveMinuteLMPsAtLocation(
  start_date = Sys.Date(),
  end_date = Sys.Date(),
  ...
)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=getFiveMinuteLMPsForDayAtLocation]{getFiveMinuteLMPsForDayAtLocation}}
  \describe{
    \item{\code{day}}{date in \code{'\%Y-\%m-\%d'}, or object coercible to class \code{Date}}
    \item{\code{user}}{user account}
    \item{\code{location_id}}{pnode ID for location}
    \item{\code{password}}{password}
    \item{\code{out.tz}}{output timezone.}
  }}
}
\description{
Get historical 5-minute LMPs for a date range
}
