% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHistorical5MinLMPs.R
\name{getFiveMinuteLMPsForDayAtLocation}
\alias{getFiveMinuteLMPsForDayAtLocation}
\title{get Five Minute System Load For Day}
\usage{
getFiveMinuteLMPsForDayAtLocation(
  day = Sys.Date(),
  user = getOption(x = "ISO_NE_USER"),
  password = getOption(x = "ISO_NE_PASSWORD"),
  location_id = 4000,
  out.tz = "America/New_York",
  ...
)
}
\arguments{
\item{day}{date in \code{'\%Y-\%m-\%d'}, or object coercible to class \code{Date}}

\item{user}{user account}

\item{password}{password}

\item{location_id}{pnode ID for location}

\item{out.tz}{output timezone.}

\item{...}{passed to httr functions.}
}
\description{
May need to add \code{config = config(ssl.verifypeer = FALSE)} if SSL cert cant be found.
}
