% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMorningReportByDay.R
\name{getMorningReportByDay}
\alias{getMorningReportByDay}
\title{Get Morning Report By Day}
\usage{
getMorningReportByDay(day = Sys.Date(), user = getOption(x = "ISO_NE_USER"),
  password = getOption(x = "ISO_NE_PASSWORD"), out.tz = "America/New_York",
  ...)
}
\arguments{
\item{day}{date}

\item{user}{user account}

\item{password}{password}

\item{out.tz}{output timezone.}

\item{...}{passed to httr functions.}
}
\description{
May need to add \code{config = httr::config(ssl.verifypeer = FALSE)} if SSL cert cant be found.
}

