% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRtHourlyDemandByLocationIdForDay.R
\name{getRtHourlyDemandByLocationIdForDay}
\alias{getRtHourlyDemandByLocationIdForDay}
\title{Get Current Realtime Hourly Demand}
\usage{
getRtHourlyDemandByLocationIdForDay(day = Sys.Date() - 3, locId = 4008,
  user = getOption(x = "ISO_NE_USER"), password = getOption(x =
  "ISO_NE_PASSWORD"), out.tz = "America/New_York", ...)
}
\arguments{
\item{day}{date in YYYY-MM-DD format, or object coercible to class \code{Date}}

\item{locId}{location id of load zone (see details)}

\item{user}{user account}

\item{password}{password}

\item{out.tz}{output timezone.}

\item{...}{passed to httr functions.}
}
\description{
May need to add \code{config(ssl.verifypeer = FALSE)} if SSL cert cant be found.
}
\details{
Table of Load Zones:
\tabular{ll}{
Location \tab Location ID \cr
.Z.MAINE \tab 4001 \cr
.Z.NEWHAMPSHIRE \tab 4002 \cr
.Z.VERMONT \tab 4003 \cr
.Z.CONNECTICUT \tab 4004 \cr
.Z.RHODEISLAND \tab 4005 \cr
.Z.SEMASS \tab 4006 \cr
.Z.WCMASS \tab 4007 \cr
.Z.NEMASSBOST \tab 4008 
}
Note that historical hourly load zone information may not be available for a few days.
}

