% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSeasonalPeakHourDataForMonth.R
\name{getSeasonalPeakHourDataForMonth}
\alias{getSeasonalPeakHourDataForMonth}
\title{Get Current Seasonal Peak Hour Data}
\usage{
getSeasonalPeakHourDataForMonth(month = format(Sys.Date(), format = "\%Y\%m"),
  user = getOption(x = "ISO_NE_USER"), password = getOption(x =
  "ISO_NE_PASSWORD"), out.tz = "America/New_York", ...)
}
\arguments{
\item{month}{month in 'YYYYMM' format}

\item{user}{user account}

\item{password}{password}

\item{out.tz}{output timezone.}

\item{...}{passed to httr functions.}
}
\description{
May need to add \code{config = httr::config(ssl.verifypeer = FALSE)} if SSL cert cant be found.
}

